Moved idl tests into a seperate file
